+++ src/pages/Auth/ResetPassword.tsx
@@ -1,40 +1,39 @@
 import React from 'react';
-import { callApiMethod, usePageConfig } from "../../core";
-import { Link } from '../../core/common';
+import { usePageConfig } from "../../core";
+import { useApi } from '../../core/context';
+import { AuthForm } from '../../forms/Layout/AuthForm';
 import { useAppContext } from '../../core/context/AppContext';
-import { PreAuthForm } from '../../forms/PreAuth/PreAuthForm';
-import { PreAuthLayout } from "../../layout";
-import { Button } from 'antd';
 import { useNavigate } from 'react-router-dom';
+import { Link } from '../../core/common';
+import { Button } from 'antd';
 
-export const ForgotPasswordPage = () => {
+export const ResetPasswordPage = () => {
   return (
-    <PreAuthLayout>
-        <ForgotPasswordForm />
-    </PreAuthLayout>
+    <ResetPasswordForm />
   );
 };
 
-export const ForgotPasswordForm = () => {
+export const ResetPasswordForm = () => {
     const navigate = useNavigate();
     const { notifySuccess, notifyError } = useAppContext();
-    const { propertiesConfig, apiConfig } = usePageConfig("/forgot-password");
+    const { propertiesConfig, apiConfig } = usePageConfig("/reset-password");
+    const { callApiMethod } = useApi();
 
     const onFinish = async (payload: any) => {
         const response: any = await callApiMethod({...apiConfig, payload});
+
         if( response.status === 200 ) {
             notifySuccess(response?.message || response?.data?.message);
-            handleResetPassword();
+            navigate('/login');
         } else {
             notifyError(response?.message || response?.error)
         }
     }
+    const handleForgotPassword = () => {
+        navigate('/forgot-password');
+    }
 
-    const handleResetPassword = () => {
-        navigate('/reset-password');
-    };
-
-    return <>{ propertiesConfig && <PreAuthForm
+    return <>{ propertiesConfig && <AuthForm
             onSubmit={onFinish}
             propertiesConfig={ propertiesConfig }
             formButtons={ ["forgotPassword" ]} 
@@ -42,16 +41,19 @@ export const ForgotPasswordForm = () => {
             <div className="PreAuthLoginActions" style={{display: 'flex' }}>
                 <Link title="Back to login?" url='/login' />
             </div>
+            
             <div className="PreAuthLoginActions" style={{display: 'flex' }}>
                 <Button 
                     type = "dashed"
                     size = "middle"
                     style = {{ width: "99%", margin:"1%" }}
-                    onClick = {handleResetPassword}
+                    onClick = {handleForgotPassword}
                 > 
-                    Has verification code? Reset password.
+                    Request new verification code?.
                 </Button>
             </div>
-        </PreAuthForm>
+            
+        </AuthForm>
     }</>
+
 }
