+++ src/pages/Auth/VerifyRegistrationPage.tsx
@@ -1,23 +1,20 @@
 import React from 'react';
-import { callApiMethod, usePageConfig } from '../../core';
+import { usePageConfig } from '../../core';
+import { useApi } from '../../core/context';
 import { Link } from '../../core/common';
 import { useAppContext } from '../../core/context/AppContext';
-import { PreAuthForm } from '../../forms/PreAuth/PreAuthForm';
-import { PreAuthLayout } from '../../layout';
+import { AuthForm } from '../../forms/Layout/AuthForm';
 import { useNavigate } from 'react-router-dom';
 
 export const VerifyRegistrationPage = () => {
-  return (
-    <PreAuthLayout>
-        <VerifyRegistrationForm />
-    </PreAuthLayout>
-  );
+  return (<VerifyRegistrationForm /> );
 };
 
 const VerifyRegistrationForm = () => {
   const navigate = useNavigate();
   const { notifySuccess, notifyError } = useAppContext();
   const { propertiesConfig, apiConfig } = usePageConfig("/verify");
+  const { callApiMethod } = useApi();
 
   const onFinish = async (payload: any) => {
     const response: any = await callApiMethod({...apiConfig, payload });
@@ -31,7 +28,7 @@ const VerifyRegistrationForm = () => {
     }
   }
 
-  return <>{ propertiesConfig && <PreAuthForm
+  return <>{ propertiesConfig && <AuthForm
             onSubmit={onFinish}
             propertiesConfig={ propertiesConfig }
             formButtons={ ["submit" ]} 
@@ -39,6 +36,6 @@ const VerifyRegistrationForm = () => {
             <div className="PreAuthLoginActions" style={{display: 'flex' }}>
                 <Link title="Back to login?" url='/login' />
             </div>
-        </PreAuthForm>
+        </AuthForm>
     }</>
 }
